handled user authorization

+++ server/controllers/group.controller.js
@@ -26,7 +26,10 @@ const groupController = {
     }),
 
     createGroup: catchAsync(async (req, res, next) => {
-        const {prefferedMonument, commonLanguage, name, creatorID} = req.body;
+        const {prefferedMonument, commonLanguage, name} = req.body;
+        const creatorID = req.user.id;
+
+        if(req.role != "tourist") return res.status(400).json({error: "You are not allowed to do this action!"});
 
         if(!commonLanguage || !name || !creatorID){
             return res.status(404).json({error: "Missing required fields!"});
@@ -44,19 +47,25 @@ const groupController = {
 
     deleteGroup: catchAsync(async (req, res, next) => {
         const groupID = req.params.id;
+        const userID = req.user.id;
+
+        //only a tourist can delete his group? or can the operator delete any group he wants??
+
         const del = await client.query(
-            "DELETE FROM TOURIST_GROUP WHERE ID = $1",
-            [groupID]
+            "DELETE FROM TOURIST_GROUP WHERE ID = $1 AND CREATORID = $2",
+            [groupID, userID]
         );
+
+        if(!del.rowCount) return res.status(400).json({error: "You are not allowed to do this action!"});
 
         return res.status(200).json({msg: "Deleted Tourist Group Successfully!"});
     }),
 
     joinGroup: catchAsync(async (req, res, next) => {
-        const {touristID} = req.body; //will replace with current logged in user
+        const touristID = req.user.id;
         const groupID = req.params.id;
 
-        if(!touristID) return res.status(404).json({error: "Missing required field!"});
+        if(req.role != "tourist") return res.status(400).json({error: "You are not allowed to do this action!"});
 
         const join = client.query(
             "INSERT INTO GROUP_MEMBERS VALUES($1, $2) RETURNING *;",
@@ -64,15 +73,12 @@ const groupController = {
         );
 
         return res.status(201).json({msg: "Joined Group Successfully!", data: join.rows});
-        
     }),
     
     leaveGroup: catchAsync(async (req, res, next) => {
-        const {touristID} = req.body; //will replace with current logged in user
+        const touristID = req.user.id;
         const groupID = req.params.id;
-
-        if(!touristID) return res.status(404).json({error: "Missing required field!"});
-
+        
         const leave = client.query(
             "DELETE FROM GROUP_MEMBERS WHERE TOURISTID = $1 AND GROUPID = $2;",
             [touristID, groupID]
